Move trait to lib, show output of test, cleanup
