+++ rest_api_tests/src/endpoint_tests/mod.rs
@@ -1,3 +1,2 @@
 pub mod item;
 mod test_utils;
-pub mod traits;
